Added internal settings feature Fixed package name Small redesign Settings Update README.rst
