+++ .gitignore
@@ -1,3 +1,4 @@
 dist
 .idea
 __pycache__
+django_app_settings.egg-info
