+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = django-app-settings
-version = 0.0.1
-description = This package provides a smart settings that searches for application settings in project level settings with provided prefix.
+name = django_app_settings
+version = 0.0.2
+description = This package provides a smart settings object that searches for app settings in project level settings with provided prefix.
 long_description = file: README.rst
 url = https://github.com/1mace1/django-app-settings
 author = Nick Belyaev
